libhb: fix which packet new_chap and scr_sequence is applied to
Due to buffers containing only partial packets, we need to save new_chap
and scr_sequence from the first packet of the frame and use those values
when we receive the last packet of the frame.

Also, change how new_chap is passed through the pipeline.  Now, new_chap
is assigned to every buffer until getting to the video encoder.  This way,
if a buffer is dropped anywhere in the decode/sync/filter chain, we
don't lose track of what chapter we are in.