reader: apply rough scr offset to timestamps
libav tries to be too smart with timestamps and
enforces unnecessary conditions.  One such condition
is that subtitle timestamps must be monotonically
increasing.  To encure this is the case, we calculate
an offset upon each SCR change that will guarantee this.
This is just a very rough SCR offset.  A fine grained
offset that maintains proper sync is calculated in sync.c