sync: send output directly to sync output fifos
since sync interleaves it's output by PTS, the stream of the incoming
buffer is mostly not the same as the stream of the outgoing buffer. This
causes a delay in the data to get to it's respective fifo until the sync
work function for the output stream is called next.  Writing directly to
the output fifo fixes this.